GET request by id (Backend)

+++ backend/src/main/java/org/acme/ProjectRepository.java
@@ -9,5 +9,5 @@ import io.quarkus.hibernate.orm.panache.PanacheRepositoryBase;
 import jakarta.enterprise.context.ApplicationScoped;
 
 @ApplicationScoped
-public class ProjectRepository implements PanacheRepositoryBase<ProjectEntity, Integer> {
+public class ProjectRepository implements PanacheRepositoryBase<ProjectEntity, Long> {
 }
